requirements.txt
pandas==2.2.3
requests==2.32.3
PyYAML==6.0.2
numpy>=2.0.0
scikit-learn>=1.5.0
h3==4.2.1
streamlit==1.44.1
pydeck==0.9.1
watchdog>=4.0.0
